+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.1.0 - 2026-04-24]
+## [1.1.0 - 2026-04-28]
 ### Added
 - Dockerfile
 - Single Spa
@@ -119,6 +119,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fetch best form templates
 - Project description services
 - Polling Service for visible bridgeheads
+- Project field row category
 
 ### Changed
 - Rename accept and reject bridgehead state buttons (authorize/revoke)